requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
psycopg2-binary
flask-restx
sqlalchemy-utils
python-dotenv
